requirements.txt
accelerate
opencv-python
pillow
pandas
numpy
scipy
pyyaml
seaborn
supervision
matplotlib
sentence_transformers
groundingdino-py
torch
torchvision
httpx
anthropic
google-genai
openai
pydantic
